Remove View and boundStore API.
Add method Cube.executeQuery()
