+++ data/cube/index.js
@@ -8,7 +8,6 @@
 export * from './Cube';
 export * from './Field';
 export * from './Query';
-export * from './View';
 
 export * from './aggregate';
 export * from './filter';
